requirements.txt
tqdm
numpy
sklearn
matplotlib
pytorch
transformers
spacy